+++ CHANGELOG.md
@@ -30,6 +30,8 @@
 
 * Fix the issue of extra space at the end of returned string in the ar2en and en2ar methods. Thanks to [Hamoud Alhoqbani](https://github.com/alhoqbani).
 
+* Add methods dd2olc, olc2dd, and volc to encode, decode, and validate location coordinates (latitude and longitude in WGS84) in the [Open Location Code](https://github.com/google/open-location-code/blob/master/docs/specification.md) format.
+
 -->
 
 ## What's new in ArPHP 5.5.2 _(release date: Jan 26, 2021)_